FathomInventory/requirements.txt
beautifulsoup4==4.13.5
greenlet==3.2.4
playwright==1.55.0
graphviz==0.20.1
pyee==13.0.0
soupsieve==2.8
typing_extensions==4.15.0
requests==2.32.3
google-api-python-client==2.84.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
